ArduinoCompat- move func def out of header
Move function definitions from arduino_compat_cc3200.h to a separate cpp
file. not sure if really necessary, but seems to work.
